Rename the compose directory to ui
