+++ app/src/main/java/com/lagradost/cloudstream3/ui/compose/settings/SettingsFragment.kt
@@ -16,14 +16,14 @@ import com.lagradost.cloudstream3.utils.UIHelper.clipboardHelper
 import com.lagradost.cloudstream3.utils.UIHelper.navigate
 import com.lagradost.cloudstream3.utils.UiImage
 import com.lagradost.cloudstream3.utils.txt
-import com.lagradost.cloudstream4.compose.components.ProfileImage
-import com.lagradost.cloudstream4.compose.screens.settings.SettingsCategory
-import com.lagradost.cloudstream4.compose.screens.settings.SettingsProfileState
-import com.lagradost.cloudstream4.compose.screens.settings.SettingsScreen
-import com.lagradost.cloudstream4.compose.screens.settings.SettingsVersionState
-import com.lagradost.cloudstream4.compose.theme.CloudStreamTheme
-import com.lagradost.cloudstream4.compose.theme.loadPrimaryColor
-import com.lagradost.cloudstream4.compose.theme.loadThemeMode
+import com.lagradost.cloudstream4.ui.components.ProfileImage
+import com.lagradost.cloudstream4.ui.screens.settings.SettingsCategory
+import com.lagradost.cloudstream4.ui.screens.settings.SettingsProfileState
+import com.lagradost.cloudstream4.ui.screens.settings.SettingsScreen
+import com.lagradost.cloudstream4.ui.screens.settings.SettingsVersionState
+import com.lagradost.cloudstream4.ui.theme.CloudStreamTheme
+import com.lagradost.cloudstream4.ui.theme.loadPrimaryColor
+import com.lagradost.cloudstream4.ui.theme.loadThemeMode
 import java.text.DateFormat
 import java.text.SimpleDateFormat
 import java.util.Date
