--- composeApp/src/commonMain/kotlin/com/lagradost/cloudstream4/compose/icons/PlayCircle.kt
@@ -1,79 +1,0 @@
-package com.lagradost.cloudstream4.compose.icons
-
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.PathFillType
-import androidx.compose.ui.graphics.SolidColor
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.StrokeJoin
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.graphics.vector.path
-import androidx.compose.ui.unit.dp
-
-@Suppress("CheckReturnValue")
-public val play_circle: ImageVector
-  get() {
-    if (_play_circle != null) {
-      return _play_circle!!
-    }
-    _play_circle =
-      ImageVector.Builder(
-          name = "play_circle",
-          defaultWidth = 24.dp,
-          defaultHeight = 24.dp,
-          viewportWidth = 24f,
-          viewportHeight = 24f,
-        )
-        .apply {
-          path(
-            fill = SolidColor(Color.Black),
-            fillAlpha = 1f,
-            stroke = null,
-            strokeAlpha = 1f,
-            strokeLineWidth = 1f,
-            strokeLineCap = StrokeCap.Butt,
-            strokeLineJoin = StrokeJoin.Bevel,
-            strokeLineMiter = 1f,
-            pathFillType = PathFillType.Companion.NonZero,
-          ) {
-            moveTo(9.5f, 16.5f)
-            lineToRelative(7f, -4.5f)
-            lineTo(9.5f, 7.5f)
-            verticalLineToRelative(9f)
-            close()
-            moveTo(12f, 22f)
-            quadTo(9.93f, 22f, 8.1f, 21.21f)
-            quadTo(6.28f, 20.43f, 4.93f, 19.08f)
-            quadTo(3.58f, 17.73f, 2.79f, 15.9f)
-            reflectiveQuadTo(2f, 12f)
-            quadTo(2f, 9.92f, 2.79f, 8.1f)
-            quadTo(3.58f, 6.27f, 4.93f, 4.93f)
-            quadTo(6.28f, 3.57f, 8.1f, 2.79f)
-            quadTo(9.93f, 2f, 12f, 2f)
-            reflectiveQuadToRelative(3.9f, 0.79f)
-            reflectiveQuadToRelative(3.17f, 2.14f)
-            quadToRelative(1.35f, 1.35f, 2.14f, 3.17f)
-            quadTo(22f, 9.92f, 22f, 12f)
-            reflectiveQuadToRelative(-0.79f, 3.9f)
-            reflectiveQuadToRelative(-2.14f, 3.17f)
-            quadToRelative(-1.35f, 1.35f, -3.17f, 2.14f)
-            reflectiveQuadTo(12f, 22f)
-            close()
-            moveToRelative(0f, -2f)
-            quadToRelative(3.35f, 0f, 5.68f, -2.32f)
-            reflectiveQuadTo(20f, 12f)
-            reflectiveQuadTo(17.68f, 6.32f)
-            reflectiveQuadTo(12f, 4f)
-            reflectiveQuadTo(6.33f, 6.32f)
-            reflectiveQuadTo(4f, 12f)
-            reflectiveQuadToRelative(2.33f, 5.68f)
-            reflectiveQuadTo(12f, 20f)
-            close()
-            moveToRelative(0f, -8f)
-            close()
-          }
-        }
-        .build()
-    return _play_circle!!
-  }
-
-private var _play_circle: ImageVector? = null
