--- composeApp/src/commonMain/kotlin/com/lagradost/cloudstream4/compose/theme/CloudStreamPrimaryColor.kt
@@ -1,28 +1,0 @@
-package com.lagradost.cloudstream4.compose.theme
-
-import androidx.compose.ui.graphics.Color
-
-enum class CloudStreamPrimaryColor(val color: Color) {
-    NORMAL(Color(0xFF3D50FA)),
-    BLUE(Color(0xFF5664B7)),
-    PURPLE(Color(0xFF6200EA)),
-    GREEN(Color(0xFF00BFA5)),
-    GREEN_APPLE(Color(0xFF48E484)),
-    RED(Color(0xFFD50000)),
-    BANANA(Color(0xFFE4D448)),
-    PARTY(Color(0xFFEA596E)),
-    PINK(Color(0xFFFF1493)),
-    CARNATION_PINK(Color(0xFFBD5DA5)),
-    MAROON(Color(0xFF451010)),
-    DARK_GREEN(Color(0xFF004500)),
-    NAVY_BLUE(Color(0xFF000080)),
-    GREY(Color(0xFF515151)),
-    WHITE(Color(0xFFFFFFFF)),
-    BROWN(Color(0xFF622C00)),
-    ORANGE(Color(0xFFCE8500)),
-    DANDELION_YELLOW(Color(0xFFF5BB00)),
-    COOL_BLUE(Color(0xFF408CAC)),
-    LAVENDER(Color(0xFF6F55AF)),
-    DYNAMIC(Color(0xFF3D50FA)),
-    DYNAMIC_TWO(Color(0xFF3D50FA)),
-}
