--- composeApp/src/commonMain/kotlin/com/lagradost/cloudstream4/compose/theme/DynamicTheme.kt
@@ -1,6 +1,0 @@
-package com.lagradost.cloudstream4.compose.theme
-
-import androidx.compose.runtime.Composable
-
-@Composable
-expect fun resolveDynamicTheme(): CloudStreamColorScheme
