+++ composeApp/src/commonMain/kotlin/com/lagradost/cloudstream4/ui/theme/CloudStreamPalette.kt
@@ -1,0 +1,58 @@
+package com.lagradost.cloudstream4.ui.theme
+
+import androidx.compose.ui.graphics.Color
+
+internal object CloudStreamPalette {
+    // Default dark (AppTheme / Black)
+    val Primary = Color(0xFF3D50FA)
+    val PrimaryDark = Color(0xFF3700B3)
+    val Ongoing = Color(0xFFF53B66)
+
+    val DarkPrimaryGrayBg = Color(0xFF2B2C30)
+    val DarkBlackBg = Color(0xFF111111)
+    val DarkIconGrayBg = Color(0xFF1C1C20)
+    val DarkBoxItemBg = Color(0xFF161616)
+    val DarkText = Color(0xFFE9EAEE)
+    val DarkGrayText = Color(0xFF9BA0A4)
+    val DarkIcon = Color(0xFF9BA0A6)
+
+    // Amoled
+    val AmoledBlack = Color(0xFF000000)
+    val AmoledNearBlack = Color(0xFF111111)
+
+    // Light
+    val LightPrimaryGrayBg = Color(0xFFF1F1F1)
+    val LightBlackBg = Color(0xFFFFFFFF)
+    val LightIconGrayBg = Color(0xFFEEEEEE)
+    val LightBoxItemBg = Color(0xFFEEEEEE)
+    val LightText = Color(0xFF202125)
+    val LightGrayText = Color(0xFF5F6267)
+    val LightIcon = Color(0xFF5F6267)
+
+    // Dracula
+    val DraculaPrimaryGrayBg = Color(0xFF414450)
+    val DraculaBlackBg = Color(0xFF282A36)
+    val DraculaIconGrayBg = Color(0xFF44475A)
+    val DraculaBoxItemBg = Color(0xFF373844)
+    val DraculaText = Color(0xFFF8F8F2)
+    val DraculaGrayText = Color(0xFF6272A4)
+    val DraculaIcon = Color(0xFF6272A4)
+
+    // Lavender Dreams
+    val LavenderPrimaryGrayBg = Color(0xFFF7EEFC)
+    val LavenderBlackBg = Color(0xFFFDF0FB)
+    val LavenderIconGrayBg = Color(0xFFB794F6)
+    val LavenderBoxItemBg = Color(0xFFF8F5FF)
+    val LavenderText = Color(0xFF2D1B47)
+    val LavenderGrayText = Color(0xFF9AB3FF)
+    val LavenderIcon = Color(0xFF7C3AED)
+
+    // Silent Blue
+    val SilentBluePrimaryGrayBg = Color(0xFF282F49)
+    val SilentBlueBlackBg = Color(0xFF151A30)
+    val SilentBlueIconGrayBg = Color(0xFF3A446A)
+    val SilentBlueBoxItemBg = Color(0xFF3A446A)
+    val SilentBlueText = Color(0xFFE0E1F3)
+    val SilentBlueGrayText = Color(0xFF7B83B0)
+    val SilentBlueIcon = Color(0xFF7B83B0)
+}
