+++ composeApp/src/commonMain/kotlin/com/lagradost/cloudstream4/ui/theme/DynamicPrimaryColor.kt
@@ -1,0 +1,10 @@
+package com.lagradost.cloudstream4.ui.theme
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+
+@Composable
+expect fun resolveDynamicPrimaryColor(): Color
+
+@Composable
+expect fun resolveDynamicSecondaryColor(): Color
