+++ composeApp/src/commonMain/kotlin/com/lagradost/cloudstream4/ui/theme/DynamicTheme.kt
@@ -1,0 +1,6 @@
+package com.lagradost.cloudstream4.ui.theme
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun resolveDynamicTheme(): CloudStreamColorScheme
