--- composeApp/src/jvmMain/kotlin/com/lagradost/cloudstream4/compose/theme/DynamicPrimaryColor.jvm.kt
@@ -1,10 +1,0 @@
-package com.lagradost.cloudstream4.compose.theme
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-
-@Composable
-actual fun resolveDynamicPrimaryColor(): Color = CloudStreamPrimaryColor.NORMAL.color
-
-@Composable
-actual fun resolveDynamicSecondaryColor(): Color = CloudStreamPrimaryColor.NORMAL.color
