--- composeApp/src/jvmMain/kotlin/com/lagradost/cloudstream4/compose/theme/DynamicTheme.jvm.kt
@@ -1,7 +1,0 @@
-package com.lagradost.cloudstream4.compose.theme
-
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
-
-@Composable
-actual fun resolveDynamicTheme(): CloudStreamColorScheme = darkScheme()
