+++ composeApp/src/jvmMain/kotlin/com/lagradost/cloudstream4/ui/theme/DynamicPrimaryColor.jvm.kt
@@ -1,0 +1,10 @@
+package com.lagradost.cloudstream4.ui.theme
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+
+@Composable
+actual fun resolveDynamicPrimaryColor(): Color = CloudStreamPrimaryColor.NORMAL.color
+
+@Composable
+actual fun resolveDynamicSecondaryColor(): Color = CloudStreamPrimaryColor.NORMAL.color
