+++ composeApp/src/jvmMain/kotlin/com/lagradost/cloudstream4/ui/theme/DynamicTheme.jvm.kt
@@ -1,0 +1,7 @@
+package com.lagradost.cloudstream4.ui.theme
+
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.graphics.Color
+
+@Composable
+actual fun resolveDynamicTheme(): CloudStreamColorScheme = darkScheme()
